Initial attempt to remove cap-std dep
