+++ crates/wasi-common/Cargo.toml
@@ -30,7 +30,7 @@ fs-set-times = { workspace = true, optional = true }
 system-interface = { workspace = true, features = ["cap_std_impls"], optional = true }
 io-lifetimes = { workspace = true, optional = true }
 # Optional, enabled by tokio feature:
-tokio = { workspace = true, features = [ "rt", "fs", "time", "io-util", "net", "io-std", "rt-multi-thread"], optional = true }
+tokio = { workspace = true, features = ["time"], optional = true }
 
 cap-std = { workspace = true, optional = true }
 
@@ -74,13 +74,7 @@ sync = [
     "dep:system-interface",
     "dep:io-lifetimes",
     "use_cap_std",
-]
-tokio = [
-    "sync",
-    "wasmtime/async",
-    "wiggle/wasmtime_async",
     "dep:tokio",
-    "use_cap_std",
 ]
 use_cap_std = ["dep:cap-std"]
 exit = [ "wasmtime", "dep:libc" ]
