+++ crates/wasi-common/src/sync/_stdio.rs
