--- crates/wasi-common/src/tokio/net.rs
@@ -1,6 +1,0 @@
-pub use super::file::TcpListener;
-pub use super::file::TcpStream;
-#[cfg(unix)]
-pub use super::file::UnixListener;
-#[cfg(unix)]
-pub use super::file::UnixStream;
