--- crates/wasi-common/src/tokio/stdio.rs
@@ -1,1 +1,0 @@
-pub use super::file::{Stderr, Stdin, Stdout, stderr, stdin, stdout};
