+++ crates/wasip1/Cargo.toml
@@ -29,7 +29,7 @@ io-extras = { workspace = true }
 windows-sys = { workspace = true, features = ["Win32_System_Threading", "Win32_Security"] }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-webrogue-wasi-common = { workspace = true, default-features = false, features = ["tokio"] }
+webrogue-wasi-common = { workspace = true, default-features = false, features = ["sync"] }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
